pyproject.toml
[project]
name = "a-share-agent"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "akshare>=1.12.0",
    "openai>=1.30.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
    "rich>=13.0.0",
    "plotly>=5.0.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0.0",
]

[tool.pdm.dev-dependencies]
dev = ["pytest>=8.0.0", "pytest-asyncio>=0.23.0"]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
